requirements.txt
# Dataframes, arrays and math
numpy~=1.26.4
pandas~=2.0.3
polars~=0.20.11
pyarrow~=15.0.0

# Plotting
matplotlib~=3.8.3
seaborn~=0.13.2

# Geospatial analysis
geopandas~=0.14.3
rasterio~=1.3.9
shapely~=2.0.3
pygeoprocessing~=2.4.0
rasterstats~=0.19.0

# Modelling
pymc~=5.10.4
numpyro~=0.13.2  # For using JAX backend with PyMC
scikit-learn~=1.4.1
statsmodels~=0.14.1

# Code formatting, typing and linting
black~=24.2.0
flake8~=7.0.0
flake8-docstrings~=1.7.0
flake8-rst-docstrings~=0.3.0
pep8-naming~=0.13.3
autopep8~=2.0.4
mypy~=1.8.0
pre-commit~=3.6.2
isort~=5.13.2
pyupgrade~=3.15.1
bandit~=1.6.0

# Notebooks
jupyter~=1.0.0
ipykernel~=6.29.3

# Other
python-box~=7.1.1
